Re-read configs when restarting the server
We introduce a new class ClangdContext that holds the language client and tracks
any resources that were allocated for it; this allows us to tear down the client
and instanciate a whole new one for restarting the server, instead of just
calling stop() and activate() on one and the same instance. This makes it
possible to change config options (e.g. the clangd path, or clangd arguments),
and have that change take effect when restarting the server.
