+++ src/semantic-highlighting.ts
@@ -5,8 +5,10 @@ import * as vscode from 'vscode';
 import * as vscodelc from 'vscode-languageclient/node';
 import * as vscodelct from 'vscode-languageserver-types';
 
+import {ClangdContext} from './clangd-context';
+
 export function activate(client: vscodelc.LanguageClient,
-                         context: vscode.ExtensionContext) {
+                         context: ClangdContext) {
   const feature = new SemanticHighlightingFeature(client, context);
   context.subscriptions.push(feature);
   client.registerFeature(feature);
@@ -65,8 +67,7 @@ export class SemanticHighlightingFeature implements vscodelc.StaticFeature {
   highlighter: Highlighter;
   // Any disposables that should be cleaned up when clangd crashes.
   private subscriptions: vscode.Disposable[] = [];
-  constructor(client: vscodelc.BaseLanguageClient,
-              context: vscode.ExtensionContext) {
+  constructor(client: vscodelc.BaseLanguageClient, context: ClangdContext) {
     context.subscriptions.push(client.onDidChangeState(({newState}) => {
       if (newState == vscodelc.State.Running) {
         // Register handler for semantic highlighting notification.
